WIP: Encapsulate list sample callback
