+++ client/plots/boxplot/view/View.ts
@@ -6,7 +6,7 @@ import type { BoxPlotDom, BoxPlotSettings } from '../BoxPlot'
 import type { LegendItemEntry } from '../viewModel/LegendDataMapper'
 import type { ViewData } from '../viewModel/ViewModel'
 import type { MassAppApi } from '#mass/types/mass'
-import type { BoxPlotInteractions } from '../BoxPlotInteractions'
+import type { BoxPlotInteractions } from '../interactions/BoxPlotInteractions'
 import { BoxPlotToolTips } from './BoxPlotToolTips'
 import { BoxPlotLabelMenu } from './BoxPlotLabelMenu'
 import { rgb } from 'd3-color'
